backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
numpy>=1.24.0
pillow>=10.0.0
python-multipart>=0.0.6
opencv-python>=4.8.0
torch>=2.0.0 --index-url https://download.pytorch.org/whl/cpu
torchvision>=0.15.0 --index-url https://download.pytorch.org/whl/cpu
